reorg, parsing game details
