+++ node-cli/index.js
@@ -1,5 +1,5 @@
 const { program } = require('commander');
-const { fetchGamesPlayed } = require('./allCommand');
+const { fetchGamesPlayed, fetchGameDetails } = require('./utils');
 
 program
     .version('1.0.0')
@@ -16,16 +16,9 @@ program
             const { username, startdate, enddate } = options;
             const plays_per_game = await fetchGamesPlayed(username, startdate, enddate);
 
-            // Look up the game rank and name for each ID
-            const game_ids = Object.keys(plays_per_game).map(Number);
-            const game_ids_str = game_ids.join(',');
+            const details = await fetchGameDetails(username, plays_per_game);
+            console.log(details)
 
-            // Print the number of plays per game
-            for (const game_id of game_ids) {
-                const play_count = plays_per_game[game_id];
-                console.log(`Game ID: ${game_id}, Plays: ${play_count}`);
-            }
-            console.log(game_ids_str)
         } catch (error) {
             console.error('Error:', error.message);
             process.exit(1);
